+++ CHANGELOG.md
@@ -12,8 +12,11 @@
 ### Changed
 
 - Several of the functions for manipulating lists of filters in the
-  `Flop.Filter` module now accept lists of maps with atom keys, lists of maps with
-  string keys, and indexed maps as produced by Phoenix HTML forms as argument.
+  `Flop.Filter` module now accept lists of maps with atom keys, lists of maps
+  with string keys, and indexed maps as produced by Phoenix HTML forms as
+  argument.
+- The `empty` and `not_empty` operators now treat empty maps as empty values on
+  map fields.
 
 ## [0.19.0] - 2023-01-15